app/requirements.txt
fastapi>=0.100.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
uvicorn>=0.15.0,<0.16.0
redis>=4.2.0,<5.0.0
fastapi-cache2>=0.1.8,<0.2.0
sqlalchemy[asyncio]>=1.4.48,<2.1
asyncpg>=0.24.0,<1.0.0
pytest==7.4.2
pytest-asyncio==0.23.7
pytest-order==1.0.1
httpx==0.24.1